requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
oracledb>=1.3.0

# Utilidades opcionales
python-dotenv>=1.0.0
